+++ back/src/main/java/com/github/oneline/onelinecourse/model/user/User.java
@@ -1,6 +1,5 @@
 package com.github.oneline.onelinecourse.model.user;
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.github.oneline.onelinecourse.model.bookmark.Bookmark;
 import com.github.oneline.onelinecourse.model.comment.Comment;
 import lombok.*;
@@ -16,7 +15,6 @@ public class User {
 
     @Id // 기본키(PK)로 지정
     @Column(name = "id")
-    // @GeneratedValue : PK 생성 규칙
     private String id;
 
     @Column(name = "email", length = 50, nullable = false)  // 테이블의 컬럼을 나타내며 선언하지 않더라고 해당 클래스의 필드는 모두 컬럼이 됨
