+++ back/src/main/java/com/github/oneline/onelinecourse/service/user/UserService.java
@@ -3,6 +3,7 @@ package com.github.oneline.onelinecourse.service.user;
 import com.github.oneline.onelinecourse.model.user.User;
 import com.github.oneline.onelinecourse.repository.user.UserRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 // 트랜잭션, 도메인 기능 간의 순서를 보장하는 Service
 @Service
@@ -15,13 +16,14 @@ public class UserService {
     }
 
     // 처음 로그인인지 확인
+    @Transactional(readOnly = true)
     public User searchUser(String userId) {
-        User user = userRepository.findById(userId)
+        return userRepository.findById(userId)
                 .orElse(null);
-        return user;
     }
 
     // 처음이라서 등록 필요
+    @Transactional
     public User createUser(User user) {
         return userRepository.save(user);
     }
